+++ src/iframeResizer.contentWindow.js
@@ -32,6 +32,7 @@
     interval = 32,
     intervalTimer = null,
     logging = false,
+    mouseEvents = false,
     msgID = '[iFrameSizer]', // Must match host page msg ID
     msgIdLen = msgID.length,
     myID = '',
@@ -189,6 +190,7 @@
     checkWidthMode()
     stopInfiniteResizingOfIFrame()
     setupPublicMethods()
+    setupMouseEvents()
     startEventListeners()
     inPageLinks = setupInPageLinks()
     sendSize('init', 'Init message from host page')
@@ -216,6 +218,7 @@
     inPageLinks.enable = undefined !== data[12] ? strBool(data[12]) : false
     resizeFrom = undefined !== data[13] ? data[13] : resizeFrom
     widthCalcMode = undefined !== data[14] ? data[14] : widthCalcMode
+    mouseEvents = undefined !== data[15] ? Boolean(data[15]) : mouseEvents
   }
 
   function depricate(key) {
@@ -644,6 +647,22 @@
     }
   }
 
+  function setupMouseEvents() {
+    if (mouseEvents !== true) return
+
+    function sendMouse(e) {
+      sendMsg(0, 0, e.type, e.screenY + ':' + e.screenX)
+    }
+
+    function addMouseListener(evt, name) {
+      log('Add event listener: ' + name)
+      addEventListener(window.document, evt, sendMouse)
+    }
+
+    addMouseListener('mouseenter', 'Mouse Enter')
+    addMouseListener('mouseleave', 'Mouse Leave')
+  }
+
   function setupPublicMethods() {
     log('Enable public methods')
 
@@ -914,7 +933,7 @@
 
     var elements = document.querySelectorAll('[' + tag + ']')
 
-    if (0 === elements.length) noTaggedElementsFound()
+    if (elements.length === 0) noTaggedElementsFound()
 
     return getMaxElement(side, elements)
   }